Added an automated way to managed Vulkan object lifetime and usage across multiple queues
Former-commit-id: eb2df3a88e6b1d2ae08fbe4f018e87ec595daebc

+++ Source/BansheeVulkanRenderAPI/Source/BsVulkanHardwareBuffer.cpp
@@ -41,7 +41,7 @@ namespace BansheeEngine
 			mAllocations[i].memory = mAllocations[i].device->allocateMemory(reqs, flags);
 		}
 
-		mSizeInBytes = reqs.size;
+		mSizeInBytes = (UINT32)reqs.size;
 	}
 
 	VulkanHardwareBuffer::~VulkanHardwareBuffer()
